Add --keep option to keep downloaded archives

+++ aqt/installer.py
@@ -199,14 +199,14 @@ class Cli:
             self.parser.print_help()
             exit(1)
 
-    def call_installer(self, qt_archives, target_dir, sevenzip):
+    def call_installer(self, qt_archives, target_dir, sevenzip, keep):
         if target_dir is None:
             base_dir = os.getcwd()
         else:
             base_dir = target_dir
         tasks = []
         for arc in qt_archives.get_archives():
-            tasks.append((arc, base_dir, sevenzip))
+            tasks.append((arc, base_dir, sevenzip, keep))
         pool = multiprocessing.Pool(self.settings.concurrency)
         pool.starmap(installer, tasks)
         pool.close()
@@ -220,6 +220,7 @@ class Cli:
         os_name = args.host
         qt_version = args.qt_version
         output_dir = args.outputdir
+        keep = args.keep
         if output_dir is None:
             base_dir = os.getcwd()
         else:
@@ -305,7 +306,7 @@ class Cli:
         except ArchiveDownloadError or ArchiveListError or NoPackageFound:
             exit(1)
         target_config = qt_archives.get_target_config()
-        self.call_installer(qt_archives, output_dir, sevenzip)
+        self.call_installer(qt_archives, output_dir, sevenzip, keep)
         if not nopatch:
             Updater.update(target_config, base_dir, self.logger)
         self.logger.info("Finished installation")
@@ -321,6 +322,7 @@ class Cli:
         os_name = args.host
         qt_version = args.qt_version
         output_dir = args.outputdir
+        keep = args.keep
         if args.base is not None:
             base = args.base + "/online/qtsdkrepository/"
         else:
@@ -373,7 +375,7 @@ class Cli:
                 exit(1)
         except ArchiveDownloadError or ArchiveListError:
             exit(1)
-        self.call_installer(srcdocexamples_archives, output_dir, sevenzip)
+        self.call_installer(srcdocexamples_archives, output_dir, sevenzip, keep)
         self.logger.info("Finished installation")
         self.logger.info(
             "Time elapsed: {time:.8f} second".format(
@@ -402,6 +404,7 @@ class Cli:
         output_dir = args.outputdir
         sevenzip = self._set_sevenzip(args)
         version = args.version
+        keep = args.keep
         if args.base is not None:
             base = args.base + "/online/qtsdkrepository/"
         else:
@@ -446,7 +449,7 @@ class Cli:
                 exit(1)
         except ArchiveDownloadError or ArchiveListError:
             exit(1)
-        self.call_installer(tool_archives, output_dir, sevenzip)
+        self.call_installer(tool_archives, output_dir, sevenzip, keep)
         self.logger.info("Finished installation")
         self.logger.info(
             "Time elapsed: {time:.8f} second".format(
@@ -516,6 +519,10 @@ class Cli:
             nargs="?",
             type=float,
             help="Specify connection timeout for download site.(default: 5 sec)",
+        )
+        subparser.add_argument(
+          '-k', '--keep', action='store_true',
+          help='Keep downloaded archive when specified, otherwise remove after install'
         )
 
     def _set_module_options(self, subparser):
@@ -651,7 +658,7 @@ class Cli:
         return args.func(args)
 
 
-def installer(qt_archive, base_dir, command, response_timeout=30):
+def installer(qt_archive, base_dir, command, keep=False, response_timeout=30):
     """
     Installer function to download archive files and extract it.
     It is called through multiprocessing.Pool()
@@ -720,7 +727,8 @@ def installer(qt_archive, base_dir, command, response_timeout=30):
                         if cpe.stderr is not None:
                             logger.error(cpe.stderr)
                         raise cpe
-    os.unlink(archive)
+    if not keep:
+        os.unlink(archive)
     logger.info(
         "Finished installation of {} in {}".format(
             archive, time.perf_counter() - start_time
